security

+++ gfp/src/main/java/gestion/finanzas/personales/gfp/service/GastoServiceImpl.java
@@ -1,11 +1,14 @@
 package gestion.finanzas.personales.gfp.service;
 
+import gestion.finanzas.personales.gfp.exception.ExceptionUserNotAuthorized;
+import gestion.finanzas.personales.gfp.model.Usuario;
 import gestion.finanzas.personales.gfp.repository.GastoRepository;
 import gestion.finanzas.personales.gfp.model.Gasto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class GastoServiceImpl implements GastoService {
@@ -19,22 +22,24 @@ public class GastoServiceImpl implements GastoService {
     }
 
     @Override
-    public List<Gasto> findByCategoriaId(Long categoriaId) {
-        return gastoRepository.findByCategoriaId(categoriaId);
+    public List<Gasto> findByCategoriaId(Long categoriaId, Usuario usuario) {
+
+        return gastoRepository.findByUsuarioIdAndCategoriaId(usuario.getId(),categoriaId);
     }
 
     @Override
-    public Gasto save(Gasto gasto) {
+    public Gasto save(Gasto gasto, Usuario usuario) {
+        gasto.setUsuario(usuario);
         return gastoRepository.save(gasto);
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Usuario usuario) {
+        Optional<Gasto> gasto = gastoRepository.findByUsuarioIdAndId(usuario.getId(),id);
+        if (gasto.isPresent() && !gasto.get().getUsuario().getId().equals(usuario.getId())){
+            throw new ExceptionUserNotAuthorized();
+        }
         gastoRepository.deleteById(id);
     }
 
-    @Override
-    public Gasto findById(Long id) {
-        return gastoRepository.findById(id).orElse(null);
-    }
 }
